refactor(plugin-workflow): clean code
