+++ packages/plugins/workflow/src/client/nodes/manual/FormBlockInitializer.tsx
@@ -35,6 +35,9 @@ function InternalFormBlockInitializer({ insert, schema, ...others }) {
             useAction: '{{ useSubmit }}',
           },
           'x-designer': 'Action.Designer',
+          'x-designer-props': {
+            type: 'record',
+          },
           'x-action': `${JOB_STATUS.RESOLVED}`,
         },
       },
